eftx_pattern_toll/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
ansys-api-edb==0.2.2
ansys-edb-core==0.2.6
ansys-pythonnet==3.1.0rc6
ansys-tools-common==0.4.3
anyio==4.12.1
arrow==1.4.0
asgiref==3.11.1
attrs==25.4.0
binaryornot==0.4.4
cabarchive==0.2.4
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
contourpy==1.3.3
cookiecutter==2.6.0
customtkinter==5.2.2
cx_Freeze==8.5.3
cycler==0.12.1
darkdetect==0.8.0
defusedxml==0.7.1
distlib==0.4.0
Django==6.0.2
ezdxf==1.4.3
fastapi==0.128.0
filelock==3.20.3
flet==0.80.2
flet-cli==0.80.2
flet-desktop==0.80.2
flet-web==0.80.2
fonttools==4.60.0
freeze-core==0.5.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lief==0.17.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
msgpack==1.1.2
nodeenv==1.10.0
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==11.3.0
platformdirs==4.9.1
pluggy==1.6.0
pre_commit==4.5.1
protobuf==6.33.5
psutil==7.2.2
pyaedt==0.25.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyedb==0.69.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.2.5
pypdf==6.1.1
PyPDF2==3.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
repath==0.9.0
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rtree==1.4.1
scikit-rf==1.10.0
scipy==1.17.0
scooby==0.11.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
starlette==0.50.0
striprtf==0.0.29
text-unidecode==1.3
toml==0.10.2
tomli_w==1.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
xmltodict==1.0.2
